requirements.txt
Cython==3.0.7
enet==0.0.0
imutils==0.5.4
numpy==1.26.3
opencv-python==4.9.0.80
RPi.GPIO==0.7.1
